scripts/requirements.txt
pandas>=1.5,<2.0
psycopg2-binary==2.9.9
requests>=2.31,<3
kaggle>=1.6,<2
python-dotenv>=1,<2
openmeteo-requests>=1,<2
requests-cache>=1,<2
retry-requests>=2,<3